+++ packages/path_provider/path_provider_android/lib/path_provider_android.dart
@@ -63,6 +63,11 @@ class PathProviderAndroid extends PathProviderPlatform {
   }
 
   @override
+  Future<String?> getApplicationCachePath() {
+    return _api.getApplicationCachePath();
+  }
+
+  @override
   Future<String?> getExternalStoragePath() {
     return _api.getExternalStoragePath();
   }