+++ packages/path_provider/path_provider_windows/example/integration_test/path_provider_test.dart
@@ -28,6 +28,12 @@ void main() {
     _verifySampleFile(result, 'applicationSupport');
   });
 
+  testWidgets('getApplicationCacheDirectory', (WidgetTester tester) async {
+    final PathProviderWindows provider = PathProviderWindows();
+    final String? result = await provider.getApplicationCachePath();
+    _verifySampleFile(result, 'applicationCache');
+  });
+
   testWidgets('getDownloadsDirectory', (WidgetTester tester) async {
     final PathProviderWindows provider = PathProviderWindows();
     final String? result = await provider.getDownloadsPath();