+++ packages/in_app_purchase/in_app_purchase/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.1.8
+
+* Updates documentation on handling subscription price changes to match Android's billing client v5.
+
 ## 3.1.7
 
 * Fixes unawaited_futures violations.