+++ packages/path_provider/path_provider_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.0
+
+* Adds getApplicationCachePath() for storing app-specific cache files.
+
 ## 2.1.7
 
 * Adds compatibility with `win32` 5.x.