+++ packages/xdg_directories/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+* Adds example app to demonstrate how to use the package.
+
 ## 1.0.1
 
 * Removes `process` dependency.